requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-autoslug==1.9.9
django-cacheops==7.0.1
django-debug-toolbar==4.1.0
django-filter==23.2
django-graphql-jwt==0.3.4
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.6
funcy==1.18
graphene==3.2.2
graphene-django==3.1.2
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
inflection==0.5.1
lxml==4.9.2
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
promise==2.3
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pypng==0.20220715.0
python-dotenv==1.0.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
